Finished the basics, need addr modes and op codes

+++ src/components/bus.rs
@@ -6,8 +6,8 @@ pub struct Bus {
 }
 
 pub trait BusActions {
-    fn write(&mut self, addr: usize, data: u8);
-    fn read(&mut self, addr: usize, read_only: bool) -> u8;
+    fn write(&mut self, addr: u16, data: u8);
+    fn read(&mut self, addr: u16, read_only: bool) -> u8;
 }
 
 impl Bus {
@@ -19,19 +19,19 @@ impl Bus {
 }
 
 impl BusActions for Bus {
-    fn read(&mut self, addr: usize, read_only: bool) -> u8 {
+    fn read(&mut self, addr: u16, read_only: bool) -> u8 {
         if addr < 0xFFFF {
             match read_only {
-                true => self.ram[addr],
-                false => self.ram[addr],
+                true => self.ram[addr as usize],
+                false => self.ram[addr as usize],
             }
         } else {
             0u8
         }
     }
-    fn write(&mut self, addr: usize, data: u8) {
+    fn write(&mut self, addr: u16, data: u8) {
         if addr < 0xFFFF {
-            self.ram[addr] = data;
+            self.ram[addr as usize] = data;
         }
     }
 }
